Move BitmapImage metadata retrieving to a new class named BitmapImageDescriptor https://bugs.webkit.org/show_bug.cgi?id=274716 rdar://128738620
Reviewed by Cameron McCormack.

This will decouple retrieving and caching the BitmapImage metadata from the
heavy weight class BitmapImageSource.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/BitmapImageDescriptor.cpp: Added.
(WebCore::BitmapImageDescriptor::BitmapImageDescriptor):
(WebCore::BitmapImageDescriptor::imageMetadata const):
(WebCore::BitmapImageDescriptor::primaryImageFrameMetadata const):
(WebCore::BitmapImageDescriptor::primaryNativeImageMetadata const):
(WebCore::BitmapImageDescriptor::encodedDataStatus const):
(WebCore::BitmapImageDescriptor::size const):
(WebCore::BitmapImageDescriptor::sourceSize const):
(WebCore::BitmapImageDescriptor::densityCorrectedSize const):
(WebCore::BitmapImageDescriptor::orientation const):
(WebCore::BitmapImageDescriptor::primaryFrameIndex const):
(WebCore::BitmapImageDescriptor::frameCount const):
(WebCore::BitmapImageDescriptor::repetitionCount const):
(WebCore::BitmapImageDescriptor::colorSpace const):
(WebCore::BitmapImageDescriptor::singlePixelSolidColor const):
(WebCore::BitmapImageDescriptor::uti const):
(WebCore::BitmapImageDescriptor::filenameExtension const):
(WebCore::BitmapImageDescriptor::accessibilityDescription const):
(WebCore::BitmapImageDescriptor::hotSpot const):
(WebCore::BitmapImageDescriptor::maximumSubsamplingLevel const):
(WebCore::BitmapImageDescriptor::subsamplingLevelForScaleFactor const):
(WebCore::BitmapImageDescriptor::shouldUseQuickLookForFullscreen const):
(WebCore::BitmapImageDescriptor::dump const):
* Source/WebCore/platform/graphics/BitmapImageDescriptor.h: Added.
(WebCore::BitmapImageDescriptor::clear):
* Source/WebCore/platform/graphics/BitmapImageSource.cpp:
(WebCore::BitmapImageSource::BitmapImageSource):
(WebCore::BitmapImageSource::dataChanged):
(WebCore::BitmapImageSource::dump const):
(WebCore::BitmapImageSource::imageMetadata const): Deleted.
(WebCore::BitmapImageSource::primaryNativeImageMetadata const): Deleted.
(WebCore::BitmapImageSource::primaryImageFrameMetadata const): Deleted.
(WebCore::BitmapImageSource::encodedDataStatus const): Deleted.
(WebCore::BitmapImageSource::size const): Deleted.
(WebCore::BitmapImageSource::sourceSize const): Deleted.
(WebCore::BitmapImageSource::densityCorrectedSize const): Deleted.
(WebCore::BitmapImageSource::orientation const): Deleted.
(WebCore::BitmapImageSource::primaryFrameIndex const): Deleted.
(WebCore::BitmapImageSource::frameCount const): Deleted.
(WebCore::BitmapImageSource::repetitionCount const): Deleted.
(WebCore::BitmapImageSource::colorSpace const): Deleted.
(WebCore::BitmapImageSource::singlePixelSolidColor const): Deleted.
(WebCore::BitmapImageSource::uti const): Deleted.
(WebCore::BitmapImageSource::filenameExtension const): Deleted.
(WebCore::BitmapImageSource::accessibilityDescription const): Deleted.
(WebCore::BitmapImageSource::hotSpot const): Deleted.
(WebCore::BitmapImageSource::maximumSubsamplingLevel const): Deleted.
(WebCore::BitmapImageSource::subsamplingLevelForScaleFactor): Deleted.
(WebCore::BitmapImageSource::shouldUseQuickLookForFullscreen const): Deleted.
* Source/WebCore/platform/graphics/BitmapImageSource.h:
(WebCore::BitmapImageSource::decoderIfExists const):
(WebCore::BitmapImageSource::frames const):
(WebCore::BitmapImageSource::repetitionCount const):
(WebCore::BitmapImageSource::encodedDataStatus const):
(WebCore::BitmapImageSource::densityCorrectedSize const):
(WebCore::BitmapImageSource::maximumSubsamplingLevel const):
* Source/WebCore/platform/graphics/nicosia/NicosiaBuffer.cpp:

Canonical link: https://commits.webkit.org/279710@main
