+++ Source/WebCore/platform/graphics/BitmapImageSource.cpp
@@ -27,6 +27,7 @@
 #include "BitmapImageSource.h"
 
 #include "BitmapImage.h"
+#include "BitmapImageDescriptor.h"
 #include "GraphicsContext.h"
 #include "ImageBuffer.h"
 #include "ImageDecoder.h"
@@ -45,6 +46,7 @@ BitmapImageSource::BitmapImageSource(BitmapImage& bitmapImage, AlphaOption alpha
     : m_bitmapImage(&bitmapImage)
     , m_alphaOption(alphaOption)
     , m_gammaAndColorProfileOption(gammaAndColorProfileOption)
+    , m_descriptor(*this)
 {
 }
 
@@ -105,7 +107,7 @@ void BitmapImageSource::encodedDataStatusChanged(EncodedDataStatus status)
 
 EncodedDataStatus BitmapImageSource::dataChanged(FragmentedSharedBuffer* data, bool allDataReceived)
 {
-    m_cachedFlags = { };
+    m_descriptor.clear();
 
     auto status = setData(data, allDataReceived);
     if (status < EncodedDataStatus::TypeAvailable)
@@ -632,212 +634,6 @@ RefPtr<NativeImage> BitmapImageSource::preTransformedNativeImageAtIndex(unsigned
     return ImageBuffer::sinkIntoNativeImage(WTFMove(buffer));
 }
 
-template<typename MetadataType>
-MetadataType BitmapImageSource::imageMetadata(MetadataType& cachedValue, const MetadataType& defaultValue, CachedFlag cachedFlag, MetadataType (ImageDecoder::*functor)() const) const
-{
-    if (m_cachedFlags.contains(cachedFlag))
-        return cachedValue;
-
-    if (!m_decoder)
-        return defaultValue;
-
-    if (!m_decoder->isSizeAvailable())
-        return defaultValue;
-
-    cachedValue = (*m_decoder.*functor)();
-    m_cachedFlags.add(cachedFlag);
-    const_cast<BitmapImageSource&>(*this).didDecodeProperties(m_decoder->bytesDecodedToDetermineProperties());
-    return cachedValue;
-}
-
-template<typename MetadataType>
-MetadataType BitmapImageSource::primaryNativeImageMetadata(MetadataType& cachedValue, const MetadataType& defaultValue, CachedFlag cachedFlag, MetadataType (NativeImage::*functor)() const) const
-{
-    if (m_cachedFlags.contains(cachedFlag))
-        return cachedValue;
-
-    RefPtr nativeImage = const_cast<BitmapImageSource&>(*this).primaryNativeImage();
-    if (!nativeImage)
-        return defaultValue;
-
-    cachedValue = (*nativeImage.*functor)();
-    m_cachedFlags.add(cachedFlag);
-    return cachedValue;
-}
-
-template<typename MetadataType>
-MetadataType BitmapImageSource::primaryImageFrameMetadata(MetadataType& cachedValue, CachedFlag cachedFlag, MetadataType (ImageFrame::*functor)() const) const
-{
-    if (m_cachedFlags.contains(cachedFlag))
-        return cachedValue;
-
-    auto& frame = const_cast<BitmapImageSource&>(*this).primaryImageFrame();
-
-    // Don't cache any unavailable frame metadata. Just return the default metadata.
-    if (!frame.hasMetadata())
-        return (frame.*functor)();
-
-    cachedValue = (frame.*functor)();
-    m_cachedFlags.add(cachedFlag);
-    return cachedValue;
-}
-
-EncodedDataStatus BitmapImageSource::encodedDataStatus() const
-{
-    return imageMetadata(m_encodedDataStatus, EncodedDataStatus::Unknown, CachedFlag::EncodedDataStatus, &ImageDecoder::encodedDataStatus);
-}
-
-IntSize BitmapImageSource::size(ImageOrientation orientation) const
-{
-    auto densityCorrectedSize = this->densityCorrectedSize();
-    if (!densityCorrectedSize)
-        return sourceSize(orientation);
-
-    if (orientation == ImageOrientation::Orientation::FromImage)
-        orientation = this->orientation();
-
-    return orientation.usesWidthAsHeight() ? densityCorrectedSize->transposedSize() : *densityCorrectedSize;
-}
-
-IntSize BitmapImageSource::sourceSize(ImageOrientation orientation) const
-{
-    IntSize size;
-
-#if !USE(CG)
-    // It's possible that we have decoded the metadata, but not frame contents yet. In that case ImageDecoder claims to
-    // have the size available, but the frame cache is empty. Return the decoder size without caching in such case.
-    if (m_decoder && m_frames.isEmpty())
-        size = m_decoder->size();
-    else
-#endif
-        size = primaryImageFrameMetadata(m_size, CachedFlag::Size, &ImageFrame::size);
-
-    if (orientation == ImageOrientation::Orientation::FromImage)
-        orientation = this->orientation();
-
-    return orientation.usesWidthAsHeight() ? size.transposedSize() : size;
-}
-
-std::optional<IntSize> BitmapImageSource::densityCorrectedSize() const
-{
-    return primaryImageFrameMetadata(m_densityCorrectedSize, CachedFlag::DensityCorrectedSize, &ImageFrame::densityCorrectedSize);
-}
-
-ImageOrientation BitmapImageSource::orientation() const
-{
-    return primaryImageFrameMetadata(m_orientation, CachedFlag::Orientation, &ImageFrame::orientation);
-}
-
-unsigned BitmapImageSource::primaryFrameIndex() const
-{
-    return imageMetadata(m_primaryFrameIndex, std::size_t(0), CachedFlag::PrimaryFrameIndex, &ImageDecoder::primaryFrameIndex);
-}
-
-unsigned BitmapImageSource::frameCount() const
-{
-    return imageMetadata(m_frameCount, std::size_t(0), CachedFlag::FrameCount, &ImageDecoder::frameCount);
-}
-
-RepetitionCount BitmapImageSource::repetitionCount() const
-{
-    return imageMetadata(m_repetitionCount, static_cast<RepetitionCount>(RepetitionCountNone), CachedFlag::RepetitionCount, &ImageDecoder::repetitionCount);
-}
-
-DestinationColorSpace BitmapImageSource::colorSpace() const
-{
-    return primaryNativeImageMetadata(m_colorSpace, DestinationColorSpace::SRGB(), CachedFlag::ColorSpace, &NativeImage::colorSpace);
-}
-
-std::optional<Color> BitmapImageSource::singlePixelSolidColor() const
-{
-    if (!hasSolidColor())
-        return std::nullopt;
-
-    return primaryNativeImageMetadata(m_singlePixelSolidColor, std::optional<Color>(), CachedFlag::SinglePixelSolidColor, &NativeImage::singlePixelSolidColor);
-}
-
-String BitmapImageSource::uti() const
-{
-#if USE(CG)
-    return imageMetadata(m_uti, String(), CachedFlag::UTI, &ImageDecoder::uti);
-#else
-    return String();
-#endif
-}
-
-String BitmapImageSource::filenameExtension() const
-{
-    return imageMetadata(m_filenameExtension, String(), CachedFlag::FilenameExtension, &ImageDecoder::filenameExtension);
-}
-
-String BitmapImageSource::accessibilityDescription() const
-{
-    return imageMetadata(m_accessibilityDescription, String(), CachedFlag::AccessibilityDescription, &ImageDecoder::accessibilityDescription);
-}
-
-std::optional<IntPoint> BitmapImageSource::hotSpot() const
-{
-    return imageMetadata(m_hotSpot, std::optional<IntPoint>(), CachedFlag::HotSpot, &ImageDecoder::hotSpot);
-}
-
-SubsamplingLevel BitmapImageSource::maximumSubsamplingLevel() const
-{
-    if (m_cachedFlags.contains(CachedFlag::MaximumSubsamplingLevel))
-        return m_maximumSubsamplingLevel;
-
-    if (!m_decoder)
-        return SubsamplingLevel::Default;
-
-    if (!m_decoder->isSizeAvailable())
-        return SubsamplingLevel::Default;
-
-    // FIXME: this value was chosen to be appropriate for iOS since the image
-    // subsampling is only enabled by default on iOS. Choose a different value
-    // if image subsampling is enabled on other platform.
-    static constexpr int maximumImageAreaBeforeSubsampling = 5 * 1024 * 1024;
-    auto level = SubsamplingLevel::First;
-
-    for (; level < SubsamplingLevel::Last; ++level) {
-        if (frameSizeAtIndex(0, level).area() < maximumImageAreaBeforeSubsampling)
-            break;
-    }
-
-    m_maximumSubsamplingLevel = level;
-    m_cachedFlags.add(CachedFlag::MaximumSubsamplingLevel);
-    return m_maximumSubsamplingLevel;
-}
-
-SubsamplingLevel BitmapImageSource::subsamplingLevelForScaleFactor(GraphicsContext& context, const FloatSize& scaleFactor, AllowImageSubsampling allowImageSubsampling)
-{
-#if USE(CG)
-    if (allowImageSubsampling == AllowImageSubsampling::No)
-        return SubsamplingLevel::Default;
-
-    // Never use subsampled images for drawing into PDF contexts.
-    if (context.hasPlatformContext() && CGContextGetType(context.platformContext()) == kCGContextTypePDF)
-        return SubsamplingLevel::Default;
-
-    float scale = std::min(float(1), std::max(scaleFactor.width(), scaleFactor.height()));
-    if (!(scale > 0 && scale <= 1))
-        return SubsamplingLevel::Default;
-
-    int result = std::ceil(std::log2(1 / scale));
-    return static_cast<SubsamplingLevel>(std::min(result, static_cast<int>(maximumSubsamplingLevel())));
-#else
-    UNUSED_PARAM(context);
-    UNUSED_PARAM(scaleFactor);
-    UNUSED_PARAM(allowImageSubsampling);
-    return SubsamplingLevel::Default;
-#endif
-}
-
-#if ENABLE(QUICKLOOK_FULLSCREEN)
-bool BitmapImageSource::shouldUseQuickLookForFullscreen() const
-{
-    return m_decoder->shouldUseQuickLookForFullscreen();
-}
-#endif
-
 IntSize BitmapImageSource::frameSizeAtIndex(unsigned index, SubsamplingLevel subsamplingLevel) const
 {
     return const_cast<BitmapImageSource&>(*this).frameAtIndexCacheIfNeeded(index, subsamplingLevel).size();
@@ -894,15 +690,7 @@ void BitmapImageSource::dump(TextStream& ts) const
     if (m_frameAnimator)
         m_frameAnimator->dump(ts);
 
-    ts.dumpProperty("size", size());
-    ts.dumpProperty("density-corrected-size", densityCorrectedSize());
-    ts.dumpProperty("primary-frame-index", primaryFrameIndex());
-    ts.dumpProperty("frame-count", frameCount());
-    ts.dumpProperty("repetition-count", repetitionCount());
-
-    ts.dumpProperty("uti", uti());
-    ts.dumpProperty("filename-extension", filenameExtension());
-    ts.dumpProperty("accessibility-description", accessibilityDescription());
+    m_descriptor.dump(ts);
 
     ts.dumpProperty("decoded-size", m_decodedSize);
     ts.dumpProperty("decode-count-for-testing", m_decodeCountForTesting);
